cart-service/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
redis==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
pytz==2023.3